refactor UI components into separate folders
