+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'pages/list_page.dart';
+import 'ui/images/images_page.dart';
 
 void main() {
   runApp(MyApp());
