+++ lib/ui/details/detailed_page.dart
